+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - select whole line by mouse triple click
+- allow pasting clipboard text in the shell view
 
 ### Bug fixes